examples/requirements_test.txt
numpy
matplotlib
scienceplots
xraylarch

